Relocate functions of feature recogntion
